requirements.txt
Flask==2.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.19.1
google-auth==2.32.0
google-generativeai==0.3.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==20.1.0
Jinja2==3.1.2
requests==2.32.3
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.4
Werkzeug==3.0.3
Flask-Cors==4.0.0
